requirements.txt
asgiref==3.4.1
bootstrap4==0.1.0
bootstraps==1.0.1
certifi==2021.10.8
charset-normalizer==2.0.10
dj-database-url==0.5.0
Django==4.0.1
django-cors-headers==3.11.0
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
mysqlclient==2.1.0
Pillow==9.0.0
psycopg2==2.9.3
pytz==2021.3
requests==2.27.1
SQLAlchemy==1.4.31
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==5.3.0
